Eliminate "organize_rows" file
