+++ nd2k/converter/__init__.py
@@ -1,4 +1,3 @@
-from .organize_rows import organize_rows
 from .combine_by_timestamp import combine_by_timestamp
 
-__all__ = ["organize_rows", "combine_by_timestamp"]
+__all__ = ["combine_by_timestamp"]
