+++ nd2k/trade.py
@@ -132,3 +132,74 @@ class PartialTrade(TradeTraits):
 
 	def complete(self) -> Trade:
 		return Trade(**vars(self))
+
+
+def build_trades(ops: list[Operation]) -> list[Trade]:
+	trades = []
+	partials: list[PartialTrade] = []
+
+	for op in ops:
+		tr = create_or_update_trade(op, partials)
+		if tr.is_completed():
+			trades.append(tr.complete())
+			partials.remove(tr)
+
+	if len(partials):
+		organize_rows_failed(partials)
+
+	return trades
+
+
+def create_or_update_trade(op: Operation, lst: list[PartialTrade]) -> PartialTrade:
+	"""
+	Attempts to fit operation into an existing partial trade.
+	If it doesn't find any match, create a new partial trade.
+	"""
+	for pt in lst:
+		if pt.fits_as_base_asset(op):
+			pt.base_asset = op
+			return pt
+
+		if pt.fits_as_quote_asset(op):
+			pt.quote_asset = op
+			return pt
+
+		if pt.fits_as_trading_fee(op):
+			pt.trading_fee = op
+			return pt
+
+	pt = PartialTrade.from_operation(op)
+	lst.append(pt)
+	return pt
+
+
+def organize_rows_failed(lst: list[PartialTrade]) -> None:
+	"""
+	If by the end of "organize_rows" we still have partial trades,
+	something is wrong and the program shouldn't proceed.
+	"""
+	error_msg = "Error! The script went through all rows in the NovaDAX CSV "
+	error_msg+= "and could not complete the following trades:\n\n"
+
+	for pt in lst:
+		base  = print_operation(pt.base_asset)
+		quote = print_operation(pt.quote_asset)
+		fee   = print_operation(pt.trading_fee)
+
+		error_msg+= f"base asset:  {base}\n"
+		error_msg+= f"quote asset: {quote}\n"
+		error_msg+= f"trading fee: {fee}\n\n"
+
+	error_msg+= "The input file may be faulty, "
+	error_msg+= "or the script misinterpreted its contents.\n"
+	error_msg+= "If you are sure the input file is correct, please open an "
+	error_msg+= "issue at https://github.com/thiagoalessio/nd2k/issues/new "
+	error_msg+= "and attach the file that caused this error.\n"
+	print(error_msg)
+	exit(1)
+
+
+def print_operation(op: Operation | None) -> str:
+	if not op:
+		return "<empty>"
+	return " | ".join([str(v) for k,v in op.__dict__.items() if k != "type"])
